pyproject.toml
[project]
name = "nutri-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django>=5.1.7",
    "pandas>=2.2.3",
    "plotly>=6.0.1",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.0",
]
